Moved message limitation (maxToken) to ChatAssistant
